move user file up
