+++ server/user.js
@@ -2,7 +2,7 @@ var mongoose = require('mongoose')
 var bcrypt   = require('bcrypt-nodejs')
 var Moment   = require('moment')
 
-var Lookups  = require('../../shared/lookups')
+var Lookups  = require('../shared/lookups')
 var Schema   = mongoose.Schema
 
 var userSchema = new Schema({
